rename: useRenderMarker -> useMarker 이름 변경
[#914]
